Testowanie i wizualizacja (jednak było dobrze).

+++ 01_jadro_wielokata_prostego/Point.java
@@ -1,25 +1,25 @@
 public class Point {
-  private double x;
-  private double y;
+  private int x;
+  private int y;
 
   public Point() {}
 
-  public Point(double x, double y) {
+  public Point(int x, int y) {
     this.x = x;
     this.y = y;
   }
 
-  public void setX(double x) {
+  public void setX(int x) {
     this.x = x;
   }
-  public double getX() {
+  public int getX() {
     return x;
   }
 
-  public void setY(double y) {
+  public void setY(int y) {
     this.y = y;
   }
-  public double getY() {
+  public int getY() {
     return y;
   }
 }
